move common geisa parts in api
